+++ Source/Main.cpp
@@ -9,8 +9,21 @@
 	#include <Tests/TestsAll.hpp>
 #endif
 
+static void SignalHandler(int sig);
+
 int main(int argc, char* argv[])
 {	
+	// Handle abnormal exists to normally end program and release resources gracefully
+	std::signal(SIGABRT, SignalHandler); // Abnormal termination triggered by abort call
+	std::signal(SIGFPE,  SignalHandler); // Floating point exception
+	std::signal(SIGILL,  SignalHandler); // Illegal instruction - invalid function image
+	std::signal(SIGINT,  SignalHandler); // Interrupt e.g console CTRL+C
+	std::signal(SIGSEGV, SignalHandler); // Segment violation
+	std::signal(SIGTERM, SignalHandler); // Software termination signal from kill
+#if defined(ENIGMA_PLATFORM_WINDOWS)
+	std::signal(SIGBREAK, SignalHandler); //  On Windows, a click on console window close button will raise SIGBREAK
+#endif
+
 	// Initialize Logger
 	Enigma::Logger::Initialize();
 	// Initialize SQLite3 Database
@@ -62,6 +75,43 @@ int main(int argc, char* argv[])
 	return exit_code;
 }
 
+/**
+*	Handles exit signals to release program resources gracefully
+*/
+static void SignalHandler(int sig)
+{
+	const auto stringify_signal = [sig]() noexcept -> const char*
+	{
+#define RET_STR(s, desc) case s: return #s ## ": " ## desc
+		switch (sig) {
+		default: return "<unknown signal>";
+			RET_STR(SIGABRT, "Abnormal termination triggered by abort call");
+			RET_STR(SIGFPE, "Floating point exception");
+			RET_STR(SIGILL, "Illegal instruction - invalid function image");
+			RET_STR(SIGINT, "Interrupt");
+			RET_STR(SIGSEGV, "Segment violation");
+			RET_STR(SIGTERM, "Software termination signal from kill");
+#if defined(ENIGMA_PLATFORM_WINDOWS)
+			RET_STR(SIGBREAK, "Ctrl-Break sequence");
+#endif  
+#undef RET_STR
+		}
+	};
+
+	if (Enigma::Logger::GetLogger()) {
+	   ENIGMA_INFO("Signal handler invoked due abnormal exit with signal: {} ({})", sig, stringify_signal());
+	   ENIGMA_INFO("Exiting gracefully...");
+	}
+
+	// Shutdown Application
+	if (Enigma::Application::GetInstance()) Enigma::Application::GetInstance()->~Application();
+	// Shutdown SQLite3 Database
+	if (Enigma::Database::GetInstance()) Enigma::Database::Shutdown();
+	// Shutdown Logger
+	if (Enigma::Logger::GetLogger()) Enigma::Logger::Shutdown();
+
+	std::exit(EXIT_SUCCESS);
+}
 
 /**
 *	Windows Entry Point (for Release & Distribution)